Move the exclusive field operators next to the programs that use them
Emptying `compute_advection_in_vertical_momentum_equation.py`,
`compute_advection_in_horizontal_momentum_equation.py` and
`compute_diagnostics_from_normal_wind.py` of their programs left 17 field
operators behind, 15 of which have no consumer outside the two merged
programs. Those 15 move into `compute_velocity_advection.py`; the first two
modules held nothing else and are deleted.

`_interpolate_to_half_levels` and `_compute_horizontal_kinetic_energy` stay in
`compute_diagnostics_from_normal_wind.py`, which is now 36 lines, because
`compute_horizontal_velocity_quantities.py` imports them.

Pure code motion: the generated C++ for a given variant is byte-identical
across this commit, and so is its build-cache key -- checked by building the
same program from both commits into two empty caches and diffing. Only GT4Py's
translation-cache key changes, since it fingerprints the program definition and
therefore where its operators are defined, so the first run after this commit
re-translates and the next one hits the cache.
