--- model/atmosphere/dycore/src/icon4py/model/atmosphere/dycore/stencils/compute_advection_in_vertical_momentum_equation.py
@@ -1,437 +1,0 @@
-# ICON4Py - ICON inspired code in Python and GT4Py
-#
-# Copyright (c) 2022-2024, ETH Zurich and MeteoSwiss
-# All rights reserved.
-#
-# Please, refer to the LICENSE file in the root directory.
-# SPDX-License-Identifier: BSD-3-Clause
-import gt4py.next as gtx
-from gt4py.next import abs, astype, broadcast, maximum, where  # noqa: A004
-from gt4py.next.experimental import concat_where
-
-from icon4py.model.atmosphere.dycore.stencils.add_extra_diffusion_for_w_con_approaching_cfl import (
-    _add_extra_diffusion_for_w_con_approaching_cfl,
-)
-from icon4py.model.atmosphere.dycore.stencils.add_interpolated_horizontal_advection_of_w import (
-    _add_interpolated_horizontal_advection_of_w,
-)
-from icon4py.model.atmosphere.dycore.stencils.compute_horizontal_advection_term_for_vertical_velocity import (
-    _compute_horizontal_advection_term_for_vertical_velocity,
-)
-from icon4py.model.atmosphere.dycore.stencils.mo_icon_interpolation_scalar_cells2verts_scalar_ri_dsl import (
-    _mo_icon_interpolation_scalar_cells2verts_scalar_ri_dsl,
-)
-from icon4py.model.common import dimension as dims, field_type_aliases as fa, type_alias as ta
-from icon4py.model.common.interpolation.stencils.interpolate_cell_field_to_half_levels import (
-    _interpolate_cell_field_to_half_levels_vp,
-)
-from icon4py.model.common.interpolation.stencils.interpolate_to_cell_center_vp import (
-    _interpolate_to_cell_center_vp,
-)
-from icon4py.model.common.type_alias import vpfloat, wpfloat
-
-
-@gtx.field_operator
-def _interpolate_contravariant_vertical_velocity_to_full_levels(
-    contravariant_corrected_w_at_cells_on_half_levels: fa.CellKHalfField[vpfloat],
-    nlev: gtx.int32,
-) -> fa.CellKField[vpfloat]:
-    # TODO(havogt): Note that `concat_where(dims.KDim == nlev-1, ...)` is currently broken
-    # because of insufficiency in the domain inference of GT4Py,
-    # see https://github.com/GridTools/gt4py/issues/2205.
-    return concat_where(
-        dims.KDim < nlev - 1,
-        vpfloat("0.5")
-        * (
-            contravariant_corrected_w_at_cells_on_half_levels(dims.KDim - 0.5)
-            + contravariant_corrected_w_at_cells_on_half_levels(dims.KDim + 0.5)
-        ),
-        vpfloat("0.5") * contravariant_corrected_w_at_cells_on_half_levels(dims.KDim - 0.5),
-    )
-
-
-@gtx.field_operator
-def _compute_horizontal_advection_of_w(
-    w: fa.CellKHalfField[ta.wpfloat],
-    tangential_wind_on_half_levels: fa.EdgeKHalfField[ta.wpfloat],
-    vn_on_half_levels: fa.EdgeKHalfField[ta.vpfloat],
-    c_intp: gtx.Field[gtx.Dims[dims.VertexDim, dims.V2CDim], ta.wpfloat],
-    inv_dual_edge_length: fa.EdgeField[ta.wpfloat],
-    inv_primal_edge_length: fa.EdgeField[ta.wpfloat],
-    tangent_orientation: fa.EdgeField[ta.wpfloat],
-) -> fa.EdgeKHalfField[ta.vpfloat]:
-    w_at_vertices = _mo_icon_interpolation_scalar_cells2verts_scalar_ri_dsl(w, c_intp)
-
-    horizontal_advection_of_w_at_edges_on_half_levels = (
-        _compute_horizontal_advection_term_for_vertical_velocity(
-            vn_on_half_levels,
-            inv_dual_edge_length,
-            w,
-            tangential_wind_on_half_levels,
-            inv_primal_edge_length,
-            tangent_orientation,
-            w_at_vertices,
-        )
-    )
-
-    return astype(horizontal_advection_of_w_at_edges_on_half_levels, vpfloat)
-
-
-@gtx.field_operator
-def _add_vertical_advection_of_w_to_advective_vertical_wind_tendency(
-    contravariant_corrected_w_at_cells_on_half_levels: fa.CellKHalfField[vpfloat],
-    w: fa.CellKHalfField[wpfloat],
-    coeff1_dwdz: fa.CellKField[vpfloat],
-    coeff2_dwdz: fa.CellKField[vpfloat],
-) -> fa.CellKHalfField[vpfloat]:
-    contravariant_corrected_w_at_cells_on_half_levels_wp = astype(
-        contravariant_corrected_w_at_cells_on_half_levels, wpfloat
-    )
-    coeff1_dwdz_at_half_levels = coeff1_dwdz(dims.KHalfDim + 0.5)
-    coeff2_dwdz_at_half_levels = coeff2_dwdz(dims.KHalfDim + 0.5)
-    coeff1_dwdz_wp, coeff2_dwdz_wp = astype(
-        (coeff1_dwdz_at_half_levels, coeff2_dwdz_at_half_levels), wpfloat
-    )
-
-    vertical_wind_advective_tendency_wp = -contravariant_corrected_w_at_cells_on_half_levels_wp * (
-        w(dims.KHalfDim - 1) * coeff1_dwdz_wp
-        - w(dims.KHalfDim + 1) * coeff2_dwdz_wp
-        + w * astype(coeff2_dwdz_at_half_levels - coeff1_dwdz_at_half_levels, wpfloat)
-    )
-    return astype(vertical_wind_advective_tendency_wp, vpfloat)
-
-
-@gtx.field_operator
-def _compute_maximum_cfl_and_clip_contravariant_vertical_velocity(
-    ddqz_z_half: fa.CellKHalfField[ta.vpfloat],
-    contravariant_corrected_w_at_cells_on_half_levels: fa.CellKHalfField[ta.vpfloat],
-    cfl_w_limit: ta.vpfloat,
-    dtime: ta.wpfloat,
-) -> tuple[
-    fa.CellKHalfField[ta.vpfloat],
-    fa.CellKHalfField[bool],
-    fa.CellKHalfField[ta.vpfloat],
-]:
-    contravariant_corrected_w_at_cells_on_half_levels_wp, ddqz_z_half_wp = astype(
-        (contravariant_corrected_w_at_cells_on_half_levels, ddqz_z_half), wpfloat
-    )
-
-    cfl_clipping = where(
-        abs(contravariant_corrected_w_at_cells_on_half_levels) > cfl_w_limit * ddqz_z_half,
-        broadcast(True, (dims.CellDim, dims.KHalfDim)),
-        False,
-    )
-
-    vertical_cfl = where(
-        cfl_clipping,
-        contravariant_corrected_w_at_cells_on_half_levels_wp * dtime / ddqz_z_half_wp,
-        broadcast(wpfloat("0.0"), (dims.CellDim, dims.KHalfDim)),
-    )
-    vertical_cfl_vp = astype(vertical_cfl, vpfloat)
-
-    contravariant_corrected_w_at_cells_on_half_levels_wp = where(
-        (cfl_clipping) & (vertical_cfl_vp < -vpfloat("0.85")),
-        astype(-vpfloat("0.85") * ddqz_z_half, wpfloat) / dtime,
-        contravariant_corrected_w_at_cells_on_half_levels_wp,
-    )
-
-    contravariant_corrected_w_at_cells_on_half_levels_wp = where(
-        (cfl_clipping) & (vertical_cfl_vp > vpfloat("0.85")),
-        astype(vpfloat("0.85") * ddqz_z_half, wpfloat) / dtime,
-        contravariant_corrected_w_at_cells_on_half_levels_wp,
-    )
-
-    return (
-        astype(contravariant_corrected_w_at_cells_on_half_levels_wp, vpfloat),
-        cfl_clipping,
-        vertical_cfl_vp,
-    )
-
-
-@gtx.field_operator
-def _compute_contravariant_corrected_w(
-    w: fa.CellKHalfField[ta.wpfloat],
-    contravariant_correction_at_cells_on_half_levels: fa.CellKHalfField[ta.vpfloat],
-) -> fa.CellKHalfField[ta.vpfloat]:
-    contravariant_corrected_w_at_cells_on_half_levels = (
-        astype(w, vpfloat) - contravariant_correction_at_cells_on_half_levels
-    )
-
-    return contravariant_corrected_w_at_cells_on_half_levels
-
-
-@gtx.field_operator
-def _compute_contravariant_corrected_w_and_cfl(
-    w: fa.CellKHalfField[ta.wpfloat],
-    contravariant_correction_at_cells_on_half_levels: fa.CellKHalfField[ta.vpfloat],
-    ddqz_z_half: fa.CellKHalfField[ta.vpfloat],
-    cfl_w_limit: ta.vpfloat,
-    dtime: ta.wpfloat,
-    nlev: gtx.int32,
-    end_index_of_damping_layer: gtx.int32,
-) -> tuple[fa.CellKHalfField[ta.vpfloat], fa.CellKHalfField[bool], fa.CellKHalfField[ta.vpfloat]]:
-    #: intermediate variable contravariant_corrected_w_at_cells_on_half_levels is originally declared as z_w_con_c in ICON
-    contravariant_corrected_w_at_cells_on_half_levels = _compute_contravariant_corrected_w(
-        w, contravariant_correction_at_cells_on_half_levels
-    )
-
-    (contravariant_corrected_w_at_cells_on_half_levels, cfl_clipping, vertical_cfl) = concat_where(
-        (dims.KHalfDim >= maximum(2, end_index_of_damping_layer - 2)) & (dims.KHalfDim < nlev - 3),
-        _compute_maximum_cfl_and_clip_contravariant_vertical_velocity(
-            ddqz_z_half=ddqz_z_half,
-            contravariant_corrected_w_at_cells_on_half_levels=contravariant_corrected_w_at_cells_on_half_levels,
-            cfl_w_limit=cfl_w_limit,
-            dtime=dtime,
-        ),
-        (
-            contravariant_corrected_w_at_cells_on_half_levels,
-            broadcast(False, (dims.CellDim, dims.KHalfDim)),
-            broadcast(vpfloat("0.0"), (dims.CellDim, dims.KHalfDim)),
-        ),
-    )
-
-    return contravariant_corrected_w_at_cells_on_half_levels, cfl_clipping, vertical_cfl
-
-
-@gtx.field_operator
-def _compute_advective_vertical_wind_tendency(
-    w: fa.CellKHalfField[ta.wpfloat],
-    horizontal_advection_of_w_at_edges_on_half_levels: fa.EdgeKHalfField[ta.wpfloat],
-    contravariant_corrected_w_at_cells_on_half_levels: fa.CellKHalfField[ta.wpfloat],
-    cfl_clipping: fa.CellKHalfField[bool],
-    coeff1_dwdz: fa.CellKField[ta.vpfloat],
-    coeff2_dwdz: fa.CellKField[ta.vpfloat],
-    e_bln_c_s: gtx.Field[gtx.Dims[dims.CellDim, dims.C2EDim], ta.wpfloat],
-    ddqz_z_half: fa.CellKHalfField[ta.vpfloat],
-    area: fa.CellField[ta.wpfloat],
-    geofac_n2s: gtx.Field[gtx.Dims[dims.CellDim, dims.C2E2CODim], ta.wpfloat],
-    owner_mask: fa.CellField[bool],
-    scalfac_exdiff: ta.wpfloat,
-    cfl_w_limit: ta.vpfloat,
-    dtime: ta.wpfloat,
-) -> fa.CellKHalfField[ta.vpfloat]:
-    vertical_wind_advective_tendency = (
-        _add_vertical_advection_of_w_to_advective_vertical_wind_tendency(
-            contravariant_corrected_w_at_cells_on_half_levels, w, coeff1_dwdz, coeff2_dwdz
-        )
-    )
-
-    vertical_wind_advective_tendency = _add_interpolated_horizontal_advection_of_w(
-        e_bln_c_s,
-        horizontal_advection_of_w_at_edges_on_half_levels,
-        vertical_wind_advective_tendency,
-    )
-
-    vertical_wind_advective_tendency = _add_extra_diffusion_for_w_con_approaching_cfl(
-        cfl_clipping,
-        owner_mask,
-        contravariant_corrected_w_at_cells_on_half_levels,
-        ddqz_z_half,
-        area,
-        geofac_n2s,
-        w,
-        vertical_wind_advective_tendency,
-        scalfac_exdiff,
-        cfl_w_limit,
-        dtime,
-    )
-
-    return vertical_wind_advective_tendency
-
-
-@gtx.field_operator
-def _compute_advection_in_corrector_vertical_momentum(
-    w: fa.CellKHalfField[ta.wpfloat],
-    tangential_wind_on_half_levels: fa.EdgeKHalfField[ta.wpfloat],
-    vn_on_half_levels: fa.EdgeKHalfField[ta.vpfloat],
-    contravariant_correction_at_cells_on_half_levels: fa.CellKHalfField[ta.vpfloat],
-    coeff1_dwdz: fa.CellKField[ta.vpfloat],
-    coeff2_dwdz: fa.CellKField[ta.vpfloat],
-    c_intp: gtx.Field[gtx.Dims[dims.VertexDim, dims.V2CDim], ta.wpfloat],
-    inv_dual_edge_length: fa.EdgeField[ta.wpfloat],
-    inv_primal_edge_length: fa.EdgeField[ta.wpfloat],
-    tangent_orientation: fa.EdgeField[ta.wpfloat],
-    e_bln_c_s: gtx.Field[gtx.Dims[dims.CellDim, dims.C2EDim], ta.wpfloat],
-    ddqz_z_half: fa.CellKHalfField[ta.vpfloat],
-    area: fa.CellField[ta.wpfloat],
-    geofac_n2s: gtx.Field[gtx.Dims[dims.CellDim, dims.C2E2CODim], ta.wpfloat],
-    owner_mask: fa.CellField[bool],
-    scalfac_exdiff: ta.wpfloat,
-    cfl_w_limit: ta.vpfloat,
-    dtime: ta.wpfloat,
-    nlev: gtx.int32,
-    end_index_of_damping_layer: gtx.int32,
-) -> tuple[fa.CellKHalfField[ta.vpfloat], fa.CellKField[ta.vpfloat], fa.CellKHalfField[ta.vpfloat]]:
-    #: intermediate variable horizontal_advection_of_w_at_edges_on_half_levels is originally declared as z_v_grad_w in ICON
-    horizontal_advection_of_w_at_edges_on_half_levels = _compute_horizontal_advection_of_w(
-        w=w,
-        tangential_wind_on_half_levels=tangential_wind_on_half_levels,
-        vn_on_half_levels=vn_on_half_levels,
-        c_intp=c_intp,
-        inv_dual_edge_length=inv_dual_edge_length,
-        inv_primal_edge_length=inv_primal_edge_length,
-        tangent_orientation=tangent_orientation,
-    )
-
-    (
-        contravariant_corrected_w_at_cells_on_half_levels,
-        cfl_clipping,
-        vertical_cfl,
-    ) = _compute_contravariant_corrected_w_and_cfl(
-        w=w,
-        contravariant_correction_at_cells_on_half_levels=contravariant_correction_at_cells_on_half_levels,
-        ddqz_z_half=ddqz_z_half,
-        cfl_w_limit=cfl_w_limit,
-        dtime=dtime,
-        nlev=nlev,
-        end_index_of_damping_layer=end_index_of_damping_layer,
-    )
-
-    vertical_wind_advective_tendency = _compute_advective_vertical_wind_tendency(
-        w=w,
-        horizontal_advection_of_w_at_edges_on_half_levels=horizontal_advection_of_w_at_edges_on_half_levels,
-        contravariant_corrected_w_at_cells_on_half_levels=contravariant_corrected_w_at_cells_on_half_levels,
-        cfl_clipping=cfl_clipping,
-        coeff1_dwdz=coeff1_dwdz,
-        coeff2_dwdz=coeff2_dwdz,
-        e_bln_c_s=e_bln_c_s,
-        ddqz_z_half=ddqz_z_half,
-        area=area,
-        geofac_n2s=geofac_n2s,
-        owner_mask=owner_mask,
-        scalfac_exdiff=scalfac_exdiff,
-        cfl_w_limit=cfl_w_limit,
-        dtime=dtime,
-    )
-
-    contravariant_corrected_w_at_cells_on_model_levels = (
-        _interpolate_contravariant_vertical_velocity_to_full_levels(
-            contravariant_corrected_w_at_cells_on_half_levels, nlev
-        )
-    )
-
-    return (
-        vertical_wind_advective_tendency,
-        contravariant_corrected_w_at_cells_on_model_levels,
-        vertical_cfl,
-    )
-
-
-@gtx.field_operator
-def _interpolate_contravariant_correction_to_cells_on_half_levels(
-    contravariant_correction_at_edges_on_model_levels: fa.EdgeKField[ta.vpfloat],
-    e_bln_c_s: gtx.Field[gtx.Dims[dims.CellDim, dims.C2EDim], ta.wpfloat],
-    wgtfac_c: fa.CellKHalfField[ta.vpfloat],
-    nflatlev: gtx.int32,
-) -> fa.CellKHalfField[ta.vpfloat]:
-    contravariant_correction_at_cells_model_levels = _interpolate_to_cell_center_vp(
-        contravariant_correction_at_edges_on_model_levels, e_bln_c_s
-    )
-    contravariant_correction_at_cells_model_levels = astype(
-        contravariant_correction_at_cells_model_levels, vpfloat
-    )
-
-    contravariant_correction_at_cells_on_half_levels = concat_where(
-        dims.KHalfDim >= nflatlev + 1,
-        _interpolate_cell_field_to_half_levels_vp(
-            wgtfac_c=wgtfac_c, interpolant=contravariant_correction_at_cells_model_levels
-        ),
-        broadcast(vpfloat("0.0"), (dims.CellDim, dims.KHalfDim)),
-    )
-
-    return contravariant_correction_at_cells_on_half_levels
-
-
-@gtx.field_operator
-def _compute_advection_in_predictor_vertical_momentum(
-    vertical_wind_advective_tendency: fa.CellKHalfField[ta.vpfloat],
-    w: fa.CellKHalfField[ta.wpfloat],
-    tangential_wind_on_half_levels: fa.EdgeKHalfField[ta.wpfloat],
-    vn_on_half_levels: fa.EdgeKHalfField[ta.vpfloat],
-    contravariant_correction_at_edges_on_model_levels: fa.EdgeKField[ta.vpfloat],
-    coeff1_dwdz: fa.CellKField[ta.vpfloat],
-    coeff2_dwdz: fa.CellKField[ta.vpfloat],
-    c_intp: gtx.Field[gtx.Dims[dims.VertexDim, dims.V2CDim], ta.wpfloat],
-    inv_dual_edge_length: fa.EdgeField[ta.wpfloat],
-    inv_primal_edge_length: fa.EdgeField[ta.wpfloat],
-    tangent_orientation: fa.EdgeField[ta.wpfloat],
-    e_bln_c_s: gtx.Field[gtx.Dims[dims.CellDim, dims.C2EDim], ta.wpfloat],
-    wgtfac_c: fa.CellKHalfField[ta.vpfloat],
-    ddqz_z_half: fa.CellKHalfField[ta.vpfloat],
-    area: fa.CellField[ta.wpfloat],
-    geofac_n2s: gtx.Field[gtx.Dims[dims.CellDim, dims.C2E2CODim], ta.wpfloat],
-    owner_mask: fa.CellField[bool],
-    scalfac_exdiff: ta.wpfloat,
-    cfl_w_limit: ta.vpfloat,
-    dtime: ta.wpfloat,
-    skip_compute_predictor_vertical_advection: bool,
-    nflatlev: gtx.int32,
-    nlev: gtx.int32,
-    end_index_of_damping_layer: gtx.int32,
-) -> tuple[
-    fa.CellKHalfField[ta.vpfloat],
-    fa.CellKHalfField[ta.vpfloat],
-    fa.CellKField[ta.vpfloat],
-    fa.CellKHalfField[ta.vpfloat],
-]:
-    contravariant_correction_at_cells_on_half_levels = _interpolate_contravariant_correction_to_cells_on_half_levels(
-        contravariant_correction_at_edges_on_model_levels=contravariant_correction_at_edges_on_model_levels,
-        e_bln_c_s=e_bln_c_s,
-        wgtfac_c=wgtfac_c,
-        nflatlev=nflatlev,
-    )
-
-    (
-        contravariant_corrected_w_at_cells_on_half_levels,
-        cfl_clipping,
-        vertical_cfl,
-    ) = _compute_contravariant_corrected_w_and_cfl(
-        w=w,
-        contravariant_correction_at_cells_on_half_levels=contravariant_correction_at_cells_on_half_levels,
-        ddqz_z_half=ddqz_z_half,
-        cfl_w_limit=cfl_w_limit,
-        dtime=dtime,
-        nlev=nlev,
-        end_index_of_damping_layer=end_index_of_damping_layer,
-    )
-
-    if not skip_compute_predictor_vertical_advection:
-        #: intermediate variable horizontal_advection_of_w_at_edges_on_half_levels is originally declared as z_v_grad_w in ICON
-        horizontal_advection_of_w_at_edges_on_half_levels = _compute_horizontal_advection_of_w(
-            w=w,
-            tangential_wind_on_half_levels=tangential_wind_on_half_levels,
-            vn_on_half_levels=vn_on_half_levels,
-            c_intp=c_intp,
-            inv_dual_edge_length=inv_dual_edge_length,
-            inv_primal_edge_length=inv_primal_edge_length,
-            tangent_orientation=tangent_orientation,
-        )
-        vertical_wind_advective_tendency = _compute_advective_vertical_wind_tendency(
-            w=w,
-            horizontal_advection_of_w_at_edges_on_half_levels=horizontal_advection_of_w_at_edges_on_half_levels,
-            contravariant_corrected_w_at_cells_on_half_levels=contravariant_corrected_w_at_cells_on_half_levels,
-            cfl_clipping=cfl_clipping,
-            coeff1_dwdz=coeff1_dwdz,
-            coeff2_dwdz=coeff2_dwdz,
-            e_bln_c_s=e_bln_c_s,
-            ddqz_z_half=ddqz_z_half,
-            area=area,
-            geofac_n2s=geofac_n2s,
-            owner_mask=owner_mask,
-            scalfac_exdiff=scalfac_exdiff,
-            cfl_w_limit=cfl_w_limit,
-            dtime=dtime,
-        )
-
-    contravariant_corrected_w_at_cells_on_model_levels = (
-        _interpolate_contravariant_vertical_velocity_to_full_levels(
-            contravariant_corrected_w_at_cells_on_half_levels, nlev
-        )
-    )
-
-    return (
-        contravariant_correction_at_cells_on_half_levels,
-        vertical_wind_advective_tendency,
-        contravariant_corrected_w_at_cells_on_model_levels,
-        vertical_cfl,
-    )
